# Modify wp-scripts to transpile node_modules