## Description
<!-- Please describe what you have changed or added -->

I've modified `wp-scripts` to transpile modules imported from `node_modules` because a significant number of packages are transpiling down to the current JS standard instead of transpiling down to ES5. Importing these packages into an application built with `wp-scripts` breaks older browsers which are still supported by WordPress e.g. IE11.

This problem arises with many of [`sindresorhus`' packages where he primarily targets NodeJS](https://github.com/sindresorhus/ama/issues/446) as encountered [here](https://github.com/WordPress/gutenberg/pull/17963#discussion_r349374328).

[Create React App](https://github.com/facebook/create-react-app/issues/1125) and Parcel have already implemented similar solutions.

Aside: [There appears to be a broader effort](https://twitter.com/_developit/status/1196393947989512193) to solve the problems around dependencies and their targeted runtimes that doesn't affect bundle size like transpiling down to the lowest common denominator (ES5) does.

## How has this been tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, tests ran to see how -->
<!-- your change affects other areas of the code, etc. -->

I created a new `wp-scripts` app, imported `hex-rgb` and checked the build output for `let` and `const`.

## Screenshots <!-- if applicable -->

## Types of changes
<!-- What types of changes does your code introduce?  -->
<!-- Bug fix (non-breaking change which fixes an issue) -->
<!-- New feature (non-breaking change which adds functionality) -->
<!-- Breaking change (fix or feature that would cause existing functionality to not work as expected) -->

No external changes.

## Checklist:
- [x] My code is tested.
- [x] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/javascript/ -->
- [x] My code follows the accessibility standards. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/accessibility-coding-standards/ -->
- [x] My code has proper inline documentation. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/javascript/ -->
- [x] I've included developer documentation if appropriate. <!-- Handbook: https://developer.wordpress.org/block-editor/ -->
- [x] I've updated all React Native files affected by any refactorings/renamings in this PR. <!-- React Native mobile Gutenberg guidelines: https://github.com/WordPress/gutenberg/blob/master/docs/contributors/native-mobile.md -->.