@jsnajdr 2019-11-28
Transpiling `node_modules` is a difficult topic with many edge cases and we (especially @blowery) spent some considerable effort in Calypso to make it right. See the discussion (that included one of CRA maintainers) in https://github.com/Automattic/wp-calypso/pull/30196. I recommend adopting the approach that is used there.

Few areas where that could be improved are:
- unfortunately, one can't simply transpile everything in `node_modules`. Some packages (like Mapbox GL) break when transpiled twice and there needs to be a whitelist. See also https://github.com/Automattic/wp-calypso/pull/30196#discussion_r251472335
- CRA and Calypso also disable the `transform-typeof-symbol` transform, for reliability and performance reasons. See also https://github.com/facebook/create-react-app/pull/5278 and https://github.com/Automattic/wp-calypso/pull/30196#discussion_r252275173