@jameslnewell 2019-11-29
> unfortunately, one can't simply transpile everything in node_modules. Some packages (like Mapbox GL) break when transpiled twice and there needs to be a whitelist. See also Automattic/wp-calypso#30196 (comment)

How does CRA manage to avoid these issues without offering any configuration?

If we need to, how about an approach like:
- default to transpiling everything
- opt-in to transpiling specific modules via configuration e.g. something like:

`package.json`
```json
{
  "wp-scripts": {
    "build": {
      "transpile-dependencies": [ "mapbox-gl" ]
    } 
  }
}
```

> CRA and Calypso also disable the transform-typeof-symbol transform, for reliability and performance reasons. See also facebook/create-react-app#5278 and Automattic/wp-calypso#30196 (comment)

Can you please review the changes 🙏