@jsnajdr 2019-11-29
> How does CRA manage to avoid these issues without offering any configuration?

That's a good question and I don't know the answer 🙂 Looking at CRA sources, it seems they transpile everything in `node_modules` without any blacklist or whitelist. Maybe @Timer knows more? Does CRA follow the approach that was recommended in https://github.com/Automattic/wp-calypso/pull/30196#discussion_r251472335 ?