@aduth 2019-12-04
Have you (or could you) measure the impact this would have on build time?

This concerns me for a few reasons, including:

- Wasted time for most modules not needing transpilation
- As already mentioned, potential for conflict in modules not expecting to be transformed this way

There's a lot to catch up on in some of the related discussions @jsnajdr mentions.

What I'd prefer most is:

- Having better awareness of where problems exist. Something like @tellthemachines's work in #18774 should help here (I believe it is based on a similar approach in Calypso).
- Avoiding packages that make our lives difficult, when possible. My understanding of this is that it's typically encouraged that packages should target the lowest common denominator in their published versions ([reference](https://github.com/rollup/rollup/wiki/pkg.module#wait-it-just-means-import-and-export--not-other-future-javascript-features)). Granted, I can see how this becomes more of a problem over time as the language evolves and what those target environments are (especially differing between Node and the browser). 
- Opt for whitelisted transpilation when necessary.