@jameslnewell 2019-12-05
@aduth For a very simple app that I created purely for testing purposes the increase in build time is <10% on both empty and primed caches. Are you able to recommend a larger repository that uses `wp-scripts` for a more realistic test? 

<details>
  <summary>Inconclusive results from an overly simple app</summary>

**Before:**

```
first run:
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  5.27s user 1.15s system 82% cpu 7.749 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  5.23s user 1.14s system 80% cpu 7.921 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  4.83s user 1.02s system 96% cpu 6.059 total
- 5.11 avg

second run: 
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  2.11s user 0.50s system 49% cpu 5.261 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  1.82s user 0.41s system 56% cpu 3.965 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  1.75s user 0.40s system 51% cpu 4.152 total
- 1.89 avg
```

After:

```
first run:
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  5.58s user 1.17s system 75% cpu 8.894 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  5.36s user 1.08s system 93% cpu 6.858 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  5.30s user 1.05s system 92% cpu 6.851 total
- 5.41 avg - 6% increase

second run:
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  2.08s user 0.50s system 54% cpu 4.699 total
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  2.01s user 0.48s system 54% cpu 4.550 tota
./node_modules/@wordpress/scripts/bin/wp-scripts.js build  2.04s user 0.50s system 50% cpu 4.980 total
- 2.04 avg - 8% increase
```

</details>