@jsnajdr 2019-12-05
Maybe @blowery or @sgomes have some node-modules-transpilation build stats from Calypso? That's a project of a similar size to Gutenberg.