@blowery 2019-12-05
@jsnajdr I don't have any stats. We _must_ transpile to keep using [some libraries](https://github.com/Automattic/wp-calypso/blob/master/webpack.config.js#L63) and support IE11 / ES5 browsers, so how much slower it makes the build isn't really relevant.