@sgomes 2019-12-05
That said, I would assume that transpiling everything would be a significant slowdown, and that's why we currently have an explicit whitelist system, right?