@blowery 2019-12-05
We've been pretty happy with a whitelist and a [post-build validation step](https://github.com/Automattic/wp-calypso/blob/master/package.json#L297) that verifies that we've caught everything that would fail a syntax check. First-party dependencies are a pretty easy to fix problem, it's the transitive stuff that will bite you. When you have a library that's ES5 but then updates to use `debug@3`, you then have to transpile debug to keep it working... As OSS projects drop IE11 support, you're going to see more and more breakage.