@blowery 2019-12-05
@sgomes It wasn't that slow from what I recall. The blocker was that it wasn't entirely safe.