@aduth 2019-12-05
> Are you able to recommend a larger repository that uses `wp-scripts` and the default `wordpress.config.js` for a more realistic test?

To be honest, I thought we were using it in _this_ project, which was my primary concern. I think a project the scale of Gutenberg (or in fact an intention to use this for Gutenberg) is a fair target, though.