@jameslnewell 2019-12-09
Regarding the impact on performance, I've copy/pasted the webpack rule into the main `gutenberg` `wordpress.config.js` file. Transpiling _all `node_modules`_ saw an increase of 78% (46s to 82s) on an empty cache and an increase of 49% (10s to 15s) on a primed cache.

<details>

#### Before

##### Empty cache
```
./node_modules/.bin/wp-scripts build  46.56s user 2.69s system 317% cpu 15.525 total
./node_modules/.bin/wp-scripts build  46.11s user 2.63s system 314% cpu 15.499 total
./node_modules/.bin/wp-scripts build  46.56s user 2.69s system 315% cpu 15.597 total
avg=46.41
```

##### Primed cache
```
./node_modules/.bin/wp-scripts build  10.48s user 1.11s system 109% cpu 10.537 total
./node_modules/.bin/wp-scripts build  10.05s user 1.12s system 105% cpu 10.575 total
./node_modules/.bin/wp-scripts build  10.15s user 1.14s system 106% cpu 10.632 total
avg=10.23
```

#### After

##### Empty cache
```
./node_modules/.bin/wp-scripts build  81.77s user 5.78s system 471% cpu 18.572 total
./node_modules/.bin/wp-scripts build  85.50s user 5.95s system 462% cpu 19.759 total
./node_modules/.bin/wp-scripts build  80.73s user 5.78s system 459% cpu 18.816 total
avg=82.66 increase=78%
```

##### Primed cache
```
./node_modules/.bin/wp-scripts build  15.29s user 2.67s system 162% cpu 11.089 total
./node_modules/.bin/wp-scripts build  15.38s user 2.63s system 161% cpu 11.167 total
./node_modules/.bin/wp-scripts build  15.18s user 2.61s system 159% cpu 11.155 total
avg=15.28 increase=49%
```
</details>

Regarding the concern around modules not transpiling correctly, I've implemented the `safelist` approach proposed in https://github.com/WordPress/gutenberg/pull/18798#issuecomment-559644989. Happy to adjust the approach as desired.